Tests: introduce new `AbstractRealConfigTestCase`
As the `Config` class contains a number of `static` properties, it is important to avoid tests influencing each other if the tests would cause changes to the value of the static properties.

This was already handled correctly in the two places were the _real_ `Config` class is being used in the tests, but that meant those test classes now contain duplicate code.

This new `AbstractRealConfigTestCase` class removes that code duplication.

This test case should be used in the rare circumstances that the _real_ `Config` class needs to be used in the tests.

In most cases, using the `ConfigDouble` class is much preferred, but for select tests for the `Config` class itself + for tests involving the `Runner` class, using the real `Config` class cannot be avoided.
