+++ tests/Core/Config/ReportWidthTest.php
@@ -10,72 +10,15 @@
 namespace PHP_CodeSniffer\Tests\Core\Config;
 
 use PHP_CodeSniffer\Config;
-use PHPUnit\Framework\TestCase;
-use ReflectionProperty;
+use PHP_CodeSniffer\Tests\Core\Config\AbstractRealConfigTestCase;
 
 /**
  * Tests for the \PHP_CodeSniffer\Config reportWidth value.
  *
  * @covers \PHP_CodeSniffer\Config::__get
  */
-final class ReportWidthTest extends TestCase
+final class ReportWidthTest extends AbstractRealConfigTestCase
 {
-
-
-    /**
-     * Set static properties in the Config class to prevent tests influencing each other.
-     *
-     * @before
-     *
-     * @return void
-     */
-    protected function cleanConfig()
-    {
-        // Set to the property's default value to clear out potentially set values from other tests.
-        self::setStaticProperty('executablePaths', []);
-
-        // Set to a usable value to circumvent Config trying to find a phpcs.xml config file.
-        self::setStaticProperty('overriddenDefaults', ['standards' => ['PSR1']]);
-
-        // Set to values which prevent the test-runner user's `CodeSniffer.conf` file
-        // from being read and influencing the tests.
-        self::setStaticProperty('configData', []);
-        self::setStaticProperty('configDataFile', '');
-
-    }//end cleanConfig()
-
-
-    /**
-     * Clean up after each finished test.
-     *
-     * @after
-     *
-     * @return void
-     */
-    protected function resetConfig()
-    {
-        $_SERVER['argv'] = [];
-
-    }//end resetConfig()
-
-
-    /**
-     * Reset the static properties in the Config class to their true defaults to prevent this class
-     * from influencing other tests.
-     *
-     * @afterClass
-     *
-     * @return void
-     */
-    public static function resetConfigToDefaults()
-    {
-        self::setStaticProperty('overriddenDefaults', []);
-        self::setStaticProperty('executablePaths', []);
-        self::setStaticProperty('configData', null);
-        self::setStaticProperty('configDataFile', null);
-        $_SERVER['argv'] = [];
-
-    }//end resetConfigToDefaults()
 
 
     /**
@@ -88,7 +31,7 @@ final class ReportWidthTest extends TestCase
      */
     public function testReportWidthDefault()
     {
-        $config = new Config();
+        $config = new Config(['--standard=PSR1']);
 
         // Can't test the exact value as "auto" will resolve differently depending on the machine running the tests.
         $this->assertTrue(is_int($config->reportWidth), 'Report width is not an integer');
@@ -112,9 +55,9 @@ final class ReportWidthTest extends TestCase
             'show_warnings'    => '0',
         ];
 
-        $this->setStaticProperty('configData', $phpCodeSnifferConfig);
+        $this->setStaticConfigProperty('configData', $phpCodeSnifferConfig);
 
-        $config = new Config();
+        $config = new Config(['--standard=PSR1']);
 
         // Can't test the exact value as "auto" will resolve differently depending on the machine running the tests.
         $this->assertTrue(is_int($config->reportWidth), 'Report width is not an integer');
@@ -139,9 +82,9 @@ final class ReportWidthTest extends TestCase
             'report_width'     => '120',
         ];
 
-        $this->setStaticProperty('configData', $phpCodeSnifferConfig);
+        $this->setStaticConfigProperty('configData', $phpCodeSnifferConfig);
 
-        $config = new Config();
+        $config = new Config(['--standard=PSR1']);
         $this->assertSame(120, $config->reportWidth);
 
     }//end testReportWidthCanBeSetFromConfFile()
@@ -159,6 +102,7 @@ final class ReportWidthTest extends TestCase
     {
         $_SERVER['argv'] = [
             'phpcs',
+            '--standard=PSR1',
             '--report-width=100',
         ];
 
@@ -180,6 +124,7 @@ final class ReportWidthTest extends TestCase
     {
         $_SERVER['argv'] = [
             'phpcs',
+            '--standard=PSR1',
             '--report-width=100',
             '--report-width=200',
         ];
@@ -209,10 +154,11 @@ final class ReportWidthTest extends TestCase
             'report_width'     => '120',
         ];
 
-        $this->setStaticProperty('configData', $phpCodeSnifferConfig);
+        $this->setStaticConfigProperty('configData', $phpCodeSnifferConfig);
 
         $cliArgs = [
             'phpcs',
+            '--standard=PSR1',
             '--report-width=180',
         ];
 
@@ -231,7 +177,7 @@ final class ReportWidthTest extends TestCase
      */
     public function testReportWidthInputHandlingForAuto()
     {
-        $config = new Config();
+        $config = new Config(['--standard=PSR1']);
         $config->reportWidth = 'auto';
 
         // Can't test the exact value as "auto" will resolve differently depending on the machine running the tests.
@@ -254,7 +200,7 @@ final class ReportWidthTest extends TestCase
      */
     public function testReportWidthInputHandling($value, $expected)
     {
-        $config = new Config();
+        $config = new Config(['--standard=PSR1']);
         $config->reportWidth = $value;
 
         $this->assertSame($expected, $config->reportWidth);
@@ -309,24 +255,6 @@ final class ReportWidthTest extends TestCase
         ];
 
     }//end dataReportWidthInputHandling()
-
-
-    /**
-     * Helper function to set a static property on the Config class.
-     *
-     * @param string $name  The name of the property to set.
-     * @param mixed  $value The value to set the property to.
-     *
-     * @return void
-     */
-    public static function setStaticProperty($name, $value)
-    {
-        $property = new ReflectionProperty('PHP_CodeSniffer\Config', $name);
-        $property->setAccessible(true);
-        $property->setValue(null, $value);
-        $property->setAccessible(false);
-
-    }//end setStaticProperty()
 
 
 }//end class
